move coalesce to stacked_pd + general clean up
